+++ packages/apps/event-schedule-app-shared/src/v7/constants/firestore-paths.ts
@@ -1,0 +1,22 @@
+/**
+ * events
+ *   |
+ *   +-- internal/values
+ *   |     |
+ *   |     +-- email
+ *   |
+ *   +-- answers
+ *         |
+ *         +- answer1
+ *         +- answer2
+ *         +- ...
+ */
+
+export const firestorePaths = {
+  events: 'events_v7',
+  answers: 'answers',
+
+  internal: 'internal',
+  values: 'values',
+  email: 'email',
+} as const;