+++ packages/apps/event-schedule-app-shared/src/v7/types/base/notification-settings.test.ts
@@ -1,0 +1,19 @@
+import {
+  fillNotificationSettings,
+  isNotificationSettings,
+  notificationSettingsDefaultValue,
+} from './notification-settings';
+
+describe('isNotificationSettings', () => {
+  test('defaultValue should be true', () => {
+    expect(isNotificationSettings(notificationSettingsDefaultValue)).toBe(true);
+  });
+});
+
+describe('fillUser', () => {
+  test('defaultValue should be true', () => {
+    expect(fillNotificationSettings({})).toStrictEqual(
+      notificationSettingsDefaultValue
+    );
+  });
+});