+++ packages/apps/event-schedule-app/src/components/pages/create-event-schedule/edit-event-schedule-ok.tsx
@@ -1,0 +1,35 @@
+import {
+  ConfirmEmailDialogStore,
+  EditEventScheduleStore,
+} from '../../../store';
+import { ConfirmEmailDialog } from '../../organisms';
+import { EditEventScheduleEmailVerified } from './edit-event-schedule-email-verified';
+
+type Props = Readonly<{
+  eventScheduleFromDb: EventSchedule;
+}>;
+
+export const EditEventScheduleOk = memoNamed<Props>(
+  'EditEventScheduleOk',
+  ({ eventScheduleFromDb }) => {
+    const editPageIsHidden = useObservableValue(
+      ConfirmEmailDialogStore.isOpen$
+    );
+
+    const emailVerified = useObservableValue(
+      EditEventScheduleStore.emailVerified$
+    );
+
+    return (
+      <>
+        {editPageIsHidden ? undefined : (
+          <EditEventScheduleEmailVerified
+            emailVerified={emailVerified}
+            eventScheduleFromDb={eventScheduleFromDb}
+          />
+        )}
+        <ConfirmEmailDialog isOpen={editPageIsHidden} />
+      </>
+    );
+  }
+);