+++ packages/apps/event-schedule-app/src/constants/notification-settings-with-email.ts
@@ -1,0 +1,8 @@
+import { notificationSettingsDefaultValue } from '@noshiro/event-schedule-app-shared';
+import type { NotificationSettingsWithEmail } from '../types';
+
+export const notificationSettingsWithEmailDefaultValue: NotificationSettingsWithEmail =
+  {
+    ...notificationSettingsDefaultValue,
+    email: '',
+  };