+++ packages/apps/event-schedule-app/src/store/fetch-state/event-list.ts
@@ -72,7 +72,7 @@ export namespace EventListFetchState {
       setRefreshButtonIsLoading(true);
 
       api.eventList
-        .get({
+        .fetch({
           filterText,
           filterOptionState,
           showAllPastDaysEvent,