+++ packages/apps/event-schedule-app/src/store/fetch-state/event-schedule.ts
@@ -29,7 +29,7 @@ export namespace EventScheduleFetchState {
     if (eventId === undefined) return;
 
     api.event
-      .get(eventId)
+      .fetch(eventId)
       .then((result) => {
         setEventScheduleResult(result);
       })